backend/requirements.txt
cohere==5.5.2
qdrant-client==1.9.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.4
python-dotenv==1.0.0
pytest==7.4.3
pandas==2.1.3
numpy==1.24.3
flake8==6.1.0
black==23.10.1
mypy==1.7.1